fix:#03 pointerを宣言。しかしpointerの値が0のまま
